Don’t allow change_table, though really should be inspecting the individual changes

+++ lib/safe_migrations/migration_ext.rb
@@ -27,6 +27,8 @@ module SafeMigrations
           case method
           when :remove_column
             raise UnsafeRemoveColumn
+          when :change_table
+            raise UnsafeChangeTable
           when :add_index
             options = args[2]
             unless (options && options[:algorithm] == :concurrently)
